предпросмотр - пробелы, shared, subviews
