+++ WeatherProjectTestOrion/Helpers/Constants/Icons.swift
@@ -8,19 +8,19 @@
 import Foundation
 
 enum Icon {
-    public static let iconButtonListBullet = "list.bullet"
-    public static let iconButtonBackToView = "chevron.left"
-    public static let iconButtonActivEditList = "checkmark"
-    public static let iconButtonNotActivEditList = "square.and.pencil"
-    public static let iconButtonAddLocation = "plus.circle.fill"
-    public static let iconButtonAdd = "plus"
-    public static let iconButtonDelete = "xmark"
-    public static let iconWind = "wind"
-    public static let iconSafari = "safari"
-    public static let iconDrop = "drop"
-    public static let iconPlusCircle = "plus.circle.fill"
-    public static let iconMinusCircle = "minus.circle.fill"
-    public static let iconCurrentLocation = "paperplane.circle.fill"
-    public static let iconButtonDeleteText = "xmark.circle.fill"
-    public static let iconMagnifyingglass = "magnifyingglass"
+    static let iconButtonListBullet = "list.bullet"
+    static let iconButtonBackToView = "chevron.left"
+    static let iconButtonActivEditList = "checkmark"
+    static let iconButtonNotActivEditList = "square.and.pencil"
+    static let iconButtonAddLocation = "plus.circle.fill"
+    static let iconButtonAdd = "plus"
+    static let iconButtonDelete = "xmark"
+    static let iconWind = "wind"
+    static let iconSafari = "safari"
+    static let iconDrop = "drop"
+    static let iconPlusCircle = "plus.circle.fill"
+    static let iconMinusCircle = "minus.circle.fill"
+    static let iconCurrentLocation = "paperplane.circle.fill"
+    static let iconButtonDeleteText = "xmark.circle.fill"
+    static let iconMagnifyingglass = "magnifyingglass"
 }
