+++ WeatherProjectTestOrion/Helpers/Localization/Localization.swift
@@ -26,12 +26,12 @@ enum Localization: String {
     // MARK: - Buttons
     case cancel = "CANCEL"
     case add = "ADD"
-  
-  case errorDenied = "CHECK AUTHORIZATION"
-  case errorRestricted = "CHECK PARENT CONTROL"
-  case locationOff = "LOCATION OFF"
-  case geocoderError = "GEOCODER ERROR"
+    case errorDenied = "CHECK AUTHORIZATION"
+    case errorRestricted = "CHECK PARENT CONTROL"
+    case locationOff = "LOCATION OFF"
+    case geocoderError = "GEOCODER ERROR"
 }
+
 extension Localization {
     var localized: String {
         return NSLocalizedString(self.rawValue, comment: "")
